// generated/prisma/models/Currency.ts
export type CurrencyModel = runtime.Types.Result.DefaultSelection<Prisma.$CurrencyPayload>
// ...
export type $CurrencyPayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "Currency"
  objects: {
    ratesAsBase: Prisma.$ExchangeRatePayload<ExtArgs>[]
    ratesAsQuote: Prisma.$ExchangeRatePayload<ExtArgs>[]
    communityRatesAsBase: Prisma.$CommunityRatePayload<ExtArgs>[]
    communityRatesAsQuote: Prisma.$CommunityRatePayload<ExtArgs>[]
  }
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    code: string
    name: string
    symbol: string | null
    isActive: boolean
    createdAt: Date
    updatedAt: Date
  }, ExtArgs["result"]["currency"]>
  composites: {}
}