// generated/prisma/models/Provider.ts
export type ProviderModel = runtime.Types.Result.DefaultSelection<Prisma.$ProviderPayload>
// ...
export type $ProviderPayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "Provider"
  objects: {
    exchangeRates: Prisma.$ExchangeRatePayload<ExtArgs>[]
  }
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    name: string
    baseUrl: string
    isActive: boolean
    priority: number
    createdAt: Date
    updatedAt: Date
  }, ExtArgs["result"]["provider"]>
  composites: {}
}