// generated/prisma/models/ExchangeRate.ts
export type ExchangeRateModel = runtime.Types.Result.DefaultSelection<Prisma.$ExchangeRatePayload>
// ...
export type $ExchangeRatePayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "ExchangeRate"
  objects: {
    baseCurrency: Prisma.$CurrencyPayload<ExtArgs>
    quoteCurrency: Prisma.$CurrencyPayload<ExtArgs>
    provider: Prisma.$ProviderPayload<ExtArgs>
  }
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    baseCurrencyId: string
    quoteCurrencyId: string
    providerId: string
    rate: runtime.Decimal
    fetchedAt: Date
    createdAt: Date
    updatedAt: Date
  }, ExtArgs["result"]["exchangeRate"]>
  composites: {}
}